move object utils and reset object utils
