+++ lisc/objects/counts.py
@@ -6,7 +6,7 @@ from collections import defaultdict
 import numpy as np
 
 from lisc.objects.base import Base
-from lisc.objects.utils import wrap, get_max_length
+from lisc.utils.base import wrap, get_max_length
 from lisc.collect import collect_counts
 from lisc.analysis.counts import (compute_normalization, compute_association_index,
                                   compute_similarity)
